Added Tour generation based on keywords chosen
When viewing a Tour Detail, the code generates a form containing checkboxes
for the Keywords available in the Tour - keywords that can be added at will.

The user can choose, submit (via GET) and get a generated tour that includes
the chosen extra topics.

This is a first attempt that needs to be improved. Currently, the tour is
generated without regard to Individual stops. That is, a Tour moves from Stop
to Stop, and each stop has its Blocks of content. Currently, all Blocks are
listed one after the other, and separate Stops are not reflected at all.